[SDK-517] Add address and tax information to user details

+++ Sources/Preferences/PropertyListPersistence.swift
@@ -122,11 +122,14 @@ class PropertyListPersistence: PreferencesPersistence {
         
         preferencesData = [:]
         
-        do {
-            try FileManager.default.removeItem(at: preferencesPath)
-        } catch {
-            Log.error(error.localizedDescription)
+        if FileManager.default.fileExists(atPath: preferencesPath.path) {
+            do {
+                try FileManager.default.removeItem(at: preferencesPath)
+            } catch {
+                Log.error(error.localizedDescription)
+            }
         }
+        
     }
     
 }
